# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "后台管理系统",
        "title": "backend-admin",
        "contact": {},
        "version": "v1"
    },
    "paths": {
...
        "/v1/sys/token/login": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "用户登录接口",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/token/logout": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "用户退出",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "用户退出",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/token/refresh": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "用户token刷新",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "用户token刷新",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/user": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "新增用户",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "新增用户",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/user/:user_id": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新用户",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除用户",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "删除用户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/user/:user_id/role": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新用户角色",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更新用户角色",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/user/list": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "用户列表",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页面",
                        "name": "page_no",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer",
                        "description": "页大小",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UserListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.AddUserReq": {
            "type": "object",
            "required": [
                "password",
                "role_id",
                "username"
            ],
            "properties": {
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "nick_name": {
                    "description": "昵称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "role_id": {
                    "description": "角色id集合",
                    "type": "array",
                    "maxItems": 200,
                    "minItems": 0,
                    "uniqueItems": true,
                    "items": {
                        "type": "integer"
                    }
                },
                "sex": {
                    "description": "性别",
                    "type": "string"
                },
                "username": {
                    "description": "用户名(唯一且不能修改)",
                    "type": "string"
                }
            }
        },
        "dto.RoleInfo": {
            "type": "object",
            "properties": {
                "role_id": {
                    "description": "角色id",
                    "type": "integer"
                },
                "role_name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
...
        "dto.SysUserList": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "createBy": {
                    "description": "创建者",
                    "type": "integer"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "nick_name": {
                    "description": "昵称",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "roles": {
                    "description": "角色集合",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RoleInfo"
                    }
                },
                "sex": {
                    "description": "性别",
                    "type": "string"
                },
                "status": {
                    "description": "状态 -1:删除 1:正常 2:停用",
                    "type": "integer"
                },
                "updateBy": {
                    "description": "更新者",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "最后更新时间",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户id",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名(唯一且不能修改)",
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateUserReq": {
            "type": "object",
            "required": [
                "role_id"
            ],
            "properties": {
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "nick_name": {
                    "description": "昵称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "role_id": {
                    "description": "角色id集合",
                    "type": "array",
                    "maxItems": 200,
                    "minItems": 0,
                    "uniqueItems": true,
                    "items": {
                        "type": "integer"
                    }
                },
                "sex": {
                    "description": "性别",
                    "type": "string"
                }
            }
        },
        "dto.UpdateUserRoleReq": {
            "type": "object",
            "properties": {
                "role_id": {
                    "description": "角色id集合",
                    "type": "array",
                    "maxItems": 200,
                    "minItems": 0,
                    "uniqueItems": true,
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "dto.UserListResp": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "数据总数",
                    "type": "integer"
                },
                "data": {
                    "description": "数据集合",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SysUserList"
                    }
                }
            }
        },
        "dto.UserLoginReq": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "用户密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名称",
                    "type": "string"
                }
            }
        },
        "dto.UserToken": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "登录token",
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Detail string `json:\"detail\"`"
                },
                "msg": {
                    "type": "string"
                },
                "ret": {
                    "type": "integer"
                }
            }
        }
    }
}